refactor(StoredItems): 컴포넌트들을 별도의 파일로 분리
