+++ src/pages/Character/components/StoredItems/PurchaseItemModal.tsx
@@ -1,0 +1,27 @@
+import type { StoreItem } from '../../types/Item';
+import ItemPurchaseModal from '../ItemModal';
+import { ModalBackdrop } from '../ItemModal.styles';
+
+function PurchaseItemModal({
+  selectedItem,
+  setSelectedItem,
+  handlePurchaseItem,
+}: {
+  selectedItem: StoreItem | null;
+  setSelectedItem: (item: StoreItem | null) => void;
+  handlePurchaseItem: (item: StoreItem) => void;
+}) {
+  const isModalOpen = !selectedItem?.isOwned;
+
+  if (!isModalOpen || !selectedItem) {
+    return null;
+  }
+
+  return (
+    <ModalBackdrop onClick={() => setSelectedItem(null)}>
+      <ItemPurchaseModal selectedItem={selectedItem} handlePurchaseItem={handlePurchaseItem} />
+    </ModalBackdrop>
+  );
+}
+
+export default PurchaseItemModal;
