+++ src/pages/Character/components/character/CharacterTab.tsx
@@ -3,7 +3,7 @@ import { TABS } from '../../constants/tab';
 import type { SelectedItem, StoreItem } from '../../types/Item';
 import type { Tab } from '../../types/tab';
 import ItemOwnedGrid from '../ItemOwnedGrid';
-import ItemStoreGrid from '../ItemStoreGrid';
+import ItemStoreGrid from '../StoredItems/StoredItems';
 import Tabs from '../Tab';
 import { ContentContainer } from './Character.styles';
